+++ packages/core/test/serialize-test.js
@@ -108,4 +108,42 @@ describe('Serialize', () => {
       expect(result).to.equal('{"element":"category"}');
     });
   });
+
+  describe('using serializeSync', () => {
+    it('errors with unknown mediaType', () => {
+      const fury = new Fury();
+      const api = new fury.minim.elements.Category();
+
+      expect(() => {
+        fury.serializeSync({ api, mediaType: 'application/unregistered' });
+      }).to.throw('Media type did not match any registered serializer!');
+    });
+
+    it('can serialize undefined `api` by creating default category', async () => {
+      const fury = new Fury();
+      fury.use({
+        name: 'json',
+        mediaTypes: ['application/json'],
+        serializeSync: ({ api, namespace }) => JSON.stringify(namespace.serialiser.serialise(api)),
+      });
+
+      expect(
+        fury.serializeSync({ api: undefined, mediaType: 'application/json' })
+      ).to.equal('{"element":"category"}');
+    });
+
+    it('can serialize with matching adapter', async () => {
+      const fury = new Fury();
+      fury.use({
+        name: 'json',
+        mediaTypes: ['application/json'],
+        serializeSync: ({ api, namespace }) => JSON.stringify(namespace.serialiser.serialise(api)),
+      });
+
+      const api = new fury.minim.elements.Category();
+
+      const result = fury.serializeSync({ api, mediaType: 'application/json' });
+      expect(result).to.equal('{"element":"category"}');
+    });
+  });
 });