+++ packages/form-serializer/test/adapter-test.js
@@ -19,7 +19,7 @@ describe('Form Serializer Adapter', () => {
       expect(adapter.mediaTypes).to.include('multipart/form-data');
     });
 
-    it('can serialize an object element', (done) => {
+    it('can serialize an object element asynchronously', (done) => {
       const element = new fury.minim.elements.Object({ message: 'Hello world' });
 
       fury.serialize({ api: element, mediaType: 'multipart/form-data' }, (error, result) => {
@@ -29,7 +29,14 @@ describe('Form Serializer Adapter', () => {
       });
     });
 
-    it('can serialize a string element', (done) => {
+    it('can serialize an object element synchronously', () => {
+      const element = new fury.minim.elements.Object({ message: 'Hello world' });
+      const result = fury.serializeSync({ api: element, mediaType: 'multipart/form-data' });
+
+      expect(result).to.equal('--BOUNDARY\r\nContent-Disposition: form-data; name="message"\r\n\r\nHello world\r\n--BOUNDARY--\r\n');
+    });
+
+    it('can serialize a string element asynchronously', (done) => {
       const element = new fury.minim.elements.String('Hello world');
 
       fury.serialize({ api: element, mediaType: 'multipart/form-data' }, (error, result) => {
@@ -37,6 +44,13 @@ describe('Form Serializer Adapter', () => {
         expect(result).to.equal('--BOUNDARY\r\nContent-Disposition: form-data; name="undefined"\r\n\r\nHello world\r\n--BOUNDARY--\r\n');
         done();
       });
+    });
+
+    it('can serialize a string element synchronously', () => {
+      const element = new fury.minim.elements.String('Hello world');
+      const result = fury.serializeSync({ api: element, mediaType: 'multipart/form-data' });
+
+      expect(result).to.equal('--BOUNDARY\r\nContent-Disposition: form-data; name="undefined"\r\n\r\nHello world\r\n--BOUNDARY--\r\n');
     });
   });
 
@@ -49,7 +63,7 @@ describe('Form Serializer Adapter', () => {
       expect(adapter.mediaTypes).to.include('application/x-www-form-urlencoded');
     });
 
-    it('can serialize an object element', (done) => {
+    it('can serialize an object element asynchronoulsy', (done) => {
       const element = new fury.minim.elements.Object({ message: 'Hello world' });
 
       fury.serialize({ api: element, mediaType: 'application/x-www-form-urlencoded' }, (error, result) => {
@@ -59,7 +73,14 @@ describe('Form Serializer Adapter', () => {
       });
     });
 
-    it('can serialize a string element', (done) => {
+    it('can serialize an object element synchronoulsy', () => {
+      const element = new fury.minim.elements.Object({ message: 'Hello world' });
+      const result = fury.serializeSync({ api: element, mediaType: 'application/x-www-form-urlencoded' });
+
+      expect(result).to.equal('message=Hello%20world');
+    });
+
+    it('can serialize a string element asynchronously', (done) => {
       const element = new fury.minim.elements.String('Hello world');
 
       fury.serialize({ api: element, mediaType: 'application/x-www-form-urlencoded' }, (error, result) => {
@@ -67,6 +88,13 @@ describe('Form Serializer Adapter', () => {
         expect(result).to.equal('undefined=Hello%20world');
         done();
       });
+    });
+
+    it('can serialize a string element synchronously', () => {
+      const element = new fury.minim.elements.String('Hello world');
+      const result = fury.serializeSync({ api: element, mediaType: 'application/x-www-form-urlencoded' });
+
+      expect(result).to.equal('undefined=Hello%20world');
     });
   });
 });