+++ packages/json-serializer/lib/adapter.js
@@ -9,4 +9,10 @@ function serialize({ api }) {
   return new Promise(resolve => resolve(serializeJSON(api)));
 }
 
-module.exports = { name, mediaTypes, serialize };
+function serializeSync({ api }) {
+  return serializeJSON(api);
+}
+
+module.exports = {
+  name, mediaTypes, serialize, serializeSync,
+};