+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # API Elements: Core
 
+## 0.2.1 (2020-08-27)
+
+### Enhancements
+
+Added `serializeSync` method to Fury.
+
 ## 0.2.0 (2020-08-05)
 
 This package updates the version of `api-elements` being used. See